# ita_root/ita_api_organization/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Exastro IT Automation API
  description: Exastro IT Automation API
  version: "2.4"
servers:
- url: "https://{host}"
  description: API endpoint
  variables:
    host:
      default: localhost
- url: "http://{host}"
  description: API endpoint
  variables:
    host:
      default: localhost
- url: "{protocol}://{host}:{port}"
  description: API endpoint (Development)
  variables:
    protocol:
      default: https
      enum:
      - http
      - https
    host:
      default: localhost
    port:
      default: "30080"
...
paths:
...
  /api/{organization_id}/workspaces/{workspace_id}/ita/user/:
    get:
      tags:
      - User Info
      description: ユーザの権限情報とWebのテーブル設定を取得する
      operationId: get_user_auth
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_8'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.user_info_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/user/menus/:
    get:
      tags:
      - User Info
      description: ユーザがアクセス可能なメニューグループ・メニューの一覧を取得する
      operationId: get_user_menus
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_9'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.user_info_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/user/menus/panels/:
    get:
      tags:
      - User Info
      description: メニューグループの画像を取得する
      operationId: get_menu_group_panels
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_10'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.user_info_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/user/table_settings/:
    post:
      tags:
      - User Info
      description: Webのテーブル設定を登録する
      operationId: post_table_settings
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: テーブル設定
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.user_info_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/user/dashboard/:
    get:
      tags:
      - User Info
      description: Dashboardのwidget設定を取得する
      operationId: get_widget_settings
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.user_info_controller
    post:
      tags:
      - User Info
      description: Dashboardのwidget設定を登録する
      operationId: post_widget_settings
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: widget設定
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.user_info_controller
...
components:
  schemas:
...
    ERROR_TEMPLATE:
      type: object
      properties:
        result:
          type: string
          example: 001-002
        message:
          type: string
...
    inline_response_200_8:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          $ref: '#/components/schemas/inline_response_200_8_data'
        message:
          type: string
          example: SUCCESS
      example:
        result: 000-00000
        data:
          user_id: id-xxxxxxxx
          user_name: User Name
          roles:
          - role1
          - role2
          - role3
          table_settings: {}
          workspaces:
          - workspace1
          - workspace2
          - workspace3
        message: SUCCESS
    inline_response_200_9:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          $ref: '#/components/schemas/inline_response_200_9_data'
        message:
          type: string
          example: SUCCESS
      example:
        result: 000-00000
        data:
          menu_groups:
          - disp_seq: 10
            id: "101"
            parent_id: ""
            menu_group_name: メニューグループ1
            menus:
            - disp_seq: 10
              id: "10101"
              menu_name: メニュー1
              menu_name_rest: menu_1
            - disp_seq: 20
              id: "10102"
              menu_name: メニュー2
              menu_name_rest: menu_2
            - disp_seq: 30
              id: "10103"
              menu_name: メニュー3
              menu_name_rest: menu_3
          - disp_seq: 20
            id: "102"
            parent_id: "101"
            menu_group_name: メニューグループ2
            menus:
            - disp_seq: 10
              id: "10104"
              menu_name: メニュー4
              menu_name_rest: menu_4
            - disp_seq: 20
              id: "10105"
              menu_name: メニュー5
              menu_name_rest: menu_5
            - disp_seq: 30
              id: "10106"
              menu_name: メニュー6
              menu_name_rest: menu_6
          - disp_seq: 30
            id: "103"
            parent_id: "101"
            menu_group_name: メニューグループ3
            menus:
            - disp_seq: 10
              id: "10107"
              menu_name: メニュー7
              menu_name_rest: menu_7
            - disp_seq: 20
              id: "10108"
              menu_name: メニュー8
              menu_name_rest: menu_8
            - disp_seq: 30
              id: "10109"
              menu_name: メニュー9
              menu_name_rest: menu_9
        message: SUCCESS
    inline_response_200_10:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: object
          additionalProperties:
            type: string
            properties:
              panelFileEncoded:
                description: メニューグループの画像ファイルをbase64エンコードした値
          description: メニューグループIDがkeyになる
        message:
          type: string
          example: SUCCESS
      example:
        result: 000-00000
        data:
          key: data
        message: SUCCESS
    inline_response_200_11:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: boolean
        message:
          type: string
      example:
        result: 000-00000
        data: true
        message: message
...
    inline_response_200_8_data:
      type: object
      properties:
        roles:
          example:
          - role1
          - role2
          - role3
        user_id:
          example: id-xxxxxxxx
        user_name:
          example: User Name
        workspaces:
          example:
          - workspace1
          - workspace2
          - workspace3
        table_settings:
          type: object
      example:
        user_id: id-xxxxxxxx
        user_name: User Name
        roles:
        - role1
        - role2
        - role3
        table_settings: {}
        workspaces:
        - workspace1
        - workspace2
        - workspace3
    inline_response_200_9_data:
      type: object
      properties:
        menu_groups:
          example:
          - disp_seq: 10
            id: "101"
            parent_id: ""
            menu_group_name: メニューグループ1
            menus:
            - disp_seq: 10
              id: "10101"
              menu_name: メニュー1
              menu_name_rest: menu_1
            - disp_seq: 20
              id: "10102"
              menu_name: メニュー2
              menu_name_rest: menu_2
            - disp_seq: 30
              id: "10103"
              menu_name: メニュー3
              menu_name_rest: menu_3
          - disp_seq: 20
            id: "102"
            parent_id: "101"
            menu_group_name: メニューグループ2
            menus:
            - disp_seq: 10
              id: "10104"
              menu_name: メニュー4
              menu_name_rest: menu_4
            - disp_seq: 20
              id: "10105"
              menu_name: メニュー5
              menu_name_rest: menu_5
            - disp_seq: 30
              id: "10106"
              menu_name: メニュー6
              menu_name_rest: menu_6
          - disp_seq: 30
            id: "103"
            parent_id: "101"
            menu_group_name: メニューグループ3
            menus:
            - disp_seq: 10
              id: "10107"
              menu_name: メニュー7
              menu_name_rest: menu_7
            - disp_seq: 20
              id: "10108"
              menu_name: メニュー8
              menu_name_rest: menu_8
            - disp_seq: 30
              id: "10109"
              menu_name: メニュー9
              menu_name_rest: menu_9
      example:
        menu_groups:
        - disp_seq: 10
          id: "101"
          parent_id: ""
          menu_group_name: メニューグループ1
          menus:
          - disp_seq: 10
            id: "10101"
            menu_name: メニュー1
            menu_name_rest: menu_1
          - disp_seq: 20
            id: "10102"
            menu_name: メニュー2
            menu_name_rest: menu_2
          - disp_seq: 30
            id: "10103"
            menu_name: メニュー3
            menu_name_rest: menu_3
        - disp_seq: 20
          id: "102"
          parent_id: "101"
          menu_group_name: メニューグループ2
          menus:
          - disp_seq: 10
            id: "10104"
            menu_name: メニュー4
            menu_name_rest: menu_4
          - disp_seq: 20
            id: "10105"
            menu_name: メニュー5
            menu_name_rest: menu_5
          - disp_seq: 30
            id: "10106"
            menu_name: メニュー6
            menu_name_rest: menu_6
        - disp_seq: 30
          id: "103"
          parent_id: "101"
          menu_group_name: メニューグループ3
          menus:
          - disp_seq: 10
            id: "10107"
            menu_name: メニュー7
            menu_name_rest: menu_7
          - disp_seq: 20
            id: "10108"
            menu_name: メニュー8
            menu_name_rest: menu_8
          - disp_seq: 30
            id: "10109"
            menu_name: メニュー9
            menu_name_rest: menu_9
...
  securitySchemes:
    BearerAuth:
      type: http
      description: Credentials or access token for API
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth
    BasicAuth:
      type: http
      description: Value converted from ID/Password in base64
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_BasicAuth